+++ CHANGES.md
@@ -1,6 +1,12 @@
 Yoyak changelog
 ===============
 
+Version 0.5.0
+-------------
+
+To be released.
+
+
 Version 0.4.0
 -------------